+++ internal/metadata.go
@@ -1,6 +1,6 @@
 package internal
 
-var Version = "0.0.2"
+var Version = "0.0.3"
 
 type Metadata struct {
     Version string