+++ types/main.go
@@ -2,10 +2,15 @@ package types
 
 type ConfigCommands struct {
 	Dir string `json:"dir"`
-	DefaultRT string `json:"defaultRT"`
 	Ignore []string `json:"ignore"`
+}
+
+type ConfigI18n struct {
+	Dir string `json:"dir"`
+	Locales []string `json:"locales"`
 }
 
 type Config struct {
 	Commands ConfigCommands `json:"commands"`
+	I18n ConfigI18n `json:"i18n"`
 }